Homestead
Set up Laravel to be developed within Homestead to allow for a
consistent development environment for any who wish to work on this
project. This allows for the distribution to be shipped with a vagrant
file, which would only require a "vagrant up" to allow for the same
development environment for any who work on the project. Next step is
creating the desired web application.